+++ tests/jonoondb_api/document_tests.cc
@@ -3,6 +3,7 @@
 #include "gtest/gtest.h"
 #include "flatbuffers/flatbuffers.h"
 #include "test_utils.h"
+#include "jonoondb_api_test_utils.h"
 #include "document.h"
 #include "document_factory.h"
 #include "index_info_impl.h"
@@ -18,11 +19,12 @@ using namespace std;
 using namespace flatbuffers;
 using namespace jonoondb_api;
 using namespace jonoondb_test;
+using namespace jonoondb_api_test;
 
 TEST(Document, Flatbuffers_GetValues_ValidBuffer) {
   string filePath = GetSchemaFilePath("tweet.bfbs");
   string schema = File::Read(filePath);
-  BufferImpl documentData = GetTweetObject();
+  BufferImpl documentData = TestUtils::GetTweetObject();
   shared_ptr<DocumentSchema>
       docSchemaPtr(DocumentSchemaFactory::CreateDocumentSchema(
       schema, SchemaType::FLAT_BUFFERS));
